discord/requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
async-timeout==4.0.3
python-dotenv==1.0.0
pytz==2023.3